backend: move account-related functions to accounts.go
